requirements.txt
Flask==2.3.2
gunicorn==20.1.0
pvporcupine==2.1.5
python-dotenv==1.0.0
requests==2.31.0
soundfile==0.12.1

